split json aggregator property tests by spec scope: validator-contract tests move to incremental_json_validator_proptest_test.go, aggregator-surface tests stay in json_aggregator_proptest_test.go
